add the Base UI logo with copy functionality
